requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
bidict==0.23.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Faker==37.1.0
fastapi==0.115.12
fastapi-socketio==0.0.10
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
passlib==1.7.4
pillow==11.2.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pytesseract==0.3.13
python-dotenv==1.1.0
python-engineio==4.12.0
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.1
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
